Divide load test to modules.
