+++ crates/papyrus_load_test/src/main.rs
@@ -7,73 +7,13 @@
 use std::env;
 use std::fs::File;
 
-use goose::goose::{GooseUser, Scenario, Transaction, TransactionError, TransactionResult};
+use goose::goose::{Scenario, Transaction};
 use goose::{scenario, transaction, util, GooseAttack};
-use serde::de::DeserializeOwned;
-use serde::{Deserialize, Serialize};
-use serde_json::json;
-type MethodResult<T> = Result<T, Box<TransactionError>>;
+use papyrus_load_test::load_tests::*;
+use serde::Serialize;
 
-async fn post_jsonrpc_request<T: DeserializeOwned>(
-    user: &mut GooseUser,
-    method: &str,
-    params: serde_json::Value,
-) -> MethodResult<T> {
-    let request = jsonrpc_request(method, params);
-    let response = user.post_json("", &request).await?.response.map_err(|e| Box::new(e.into()))?;
-    #[derive(Deserialize)]
-    struct TransactionReceiptResponse<T> {
-        result: T,
-    }
-    let response: TransactionReceiptResponse<T> =
-        response.json().await.map_err(|e| Box::new(e.into()))?;
-
-    Ok(response.result)
-}
-
-fn jsonrpc_request(method: &str, params: serde_json::Value) -> serde_json::Value {
-    json!({
-        "jsonrpc": "2.0",
-        "id": "0",
-        "method": method,
-        "params": params,
-    })
-}
-
-async fn loadtest_get_block_with_tx_hashes_by_number(user: &mut GooseUser) -> TransactionResult {
-    let _: serde_json::Value = get_block_with_tx_hashes_by_number(user, 1).await?;
-    Ok(())
-}
-
-pub async fn get_block_with_tx_hashes_by_number<T: DeserializeOwned>(
-    user: &mut GooseUser,
-    number: u64,
-) -> MethodResult<T> {
-    post_jsonrpc_request(user, "starknet_getBlockWithTxHashes", json!([{ "block_number": number }]))
-        .await
-}
-
-async fn loadtest_get_block_with_tx_hashes_by_hash(user: &mut GooseUser) -> TransactionResult {
-    // TODO(shahak): Get a hash by getting a block instead of relying on that this hash exists.
-    let _: serde_json::Value = get_block_with_tx_hashes_by_hash(
-        user,
-        "0x1d997fd79d81bb4c30c78d7cb32fb8a59112eeb86347446235cead6194aed07",
-    )
-    .await?;
-    Ok(())
-}
-
-pub async fn get_block_with_tx_hashes_by_hash<T: DeserializeOwned>(
-    user: &mut GooseUser,
-    hash_str: &str,
-) -> MethodResult<T> {
-    post_jsonrpc_request(user, "starknet_getBlockWithTxHashes", json!([{ "block_hash": hash_str }]))
-        .await
-}
-
-#[tokio::main]
-async fn main() -> anyhow::Result<()> {
-    let metrics = GooseAttack::initialize()?
+fn register_scenarios(goose: GooseAttack) -> GooseAttack {
+    goose
         .register_scenario(
             scenario!("block_by_number")
                 .register_transaction(transaction!(loadtest_get_block_with_tx_hashes_by_number)),
@@ -82,8 +22,12 @@ async fn main() -> anyhow::Result<()> {
             scenario!("block_by_hash")
                 .register_transaction(transaction!(loadtest_get_block_with_tx_hashes_by_hash)),
         )
-        .execute()
-        .await?;
+}
+
+#[tokio::main]
+async fn main() -> anyhow::Result<()> {
+    let goose = register_scenarios(GooseAttack::initialize()?);
+    let metrics = goose.execute().await?;
 
     // The OUTPUT_FILE env is expected to be a valid path in the os.
     // If exists, aggregated results will be written to that path in the following json format:
